requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
attrs==25.1.0
bcrypt==3.2.2
beautifulsoup4==4.13.3
billiard==4.2.1
black==25.1.0
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.12
cryptography==44.0.1
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
duckdb==1.2.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.8
filetype==1.2.0
flake8==7.1.2
frozenlist==1.5.0
fsspec==2025.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.3.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
isort==6.0.0
jiter==0.8.2
joblib==1.4.2
kombu==5.4.2
llama-cloud==0.1.13
llama-cloud-services==0.6.1
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.0
llama-index-core==0.12.19
llama-index-embeddings-gemini==0.3.1
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.8
llama-index-llms-openai==0.3.20
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.5
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-pinecone==0.4.2
llama-parse==0.6.1
Mako==1.3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
mccabe==0.7.0
multidict==6.1.0
mypy==1.15.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.3
openai==1.64.0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pathspec==0.12.1
pillow==11.1.0
pinecone==6.0.1
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.0
protobuf==4.25.6
psycopg2-binary==2.9.10
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
pyflakes==3.2.0
pyparsing==3.2.1
pypdf==5.3.0
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
python-jose==3.4.0
python-magic==0.4.27
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rsa==4.9
sendgrid==6.11.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
starkbank-ecdsa==2.2.0
starlette==0.45.3
striprtf==0.0.26
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
websockets==14.2
wrapt==1.17.2
yarl==1.18.3
